% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-snippet.R
\name{add_snippet}
\alias{add_snippet}
\title{add a snippet to a document container}
\usage{
add_snippet(doc, snippet, name, ...)
}
\arguments{
\item{doc}{the document container}

\item{snippet}{the snippet from rtolatex().}

\item{name}{name of the snippet}

\item{...}{ignored only for syntax}
}
\value{
the document container with updates
}
\description{
inserts a latex snippet into the document
}
\examples{
doc <- document_container()
doc <- add_snippet(doc, rtolatex(0.94, name = "accuracy"), "accuracy")
}
