% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_container.R
\name{document_container}
\alias{document_container}
\title{creating a latex document container for all snippets}
\usage{
document_container(path = "outputs")
}
\arguments{
\item{path}{folder where all snippets are stored}
}
\value{
a document container
}
\description{
creates empty containers to store all latex snippets
use add_snippet() to add snippets
use write_textfile() to save as files
}
\examples{
doc <- document_container()
doc <- document_container("my_outputs")
}
