% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_container.R
\name{print.document_container}
\alias{print.document_container}
\title{printing document container}
\usage{
\method{print}{document_container}(x, ...)
}
\arguments{
\item{x}{the document container}

\item{...}{ignored only for syntax}
}
\description{
printing document container
}
