% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.R
\name{print.latex_snippet}
\alias{print.latex_snippet}
\title{print latex objects}
\usage{
\method{print}{latex_snippet}(x, ...)
}
\arguments{
\item{x}{a latex snippet}

\item{...}{ignored only for syntax}
}
\description{
print latex objects
}
