% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.R
\name{rtolatex.default}
\alias{rtolatex.default}
\title{alternative function when the object is not supported}
\usage{
\method{rtolatex}{default}(x, ...)
}
\arguments{
\item{x}{object that is not supported}

\item{...}{ignored only for syntax}
}
\description{
alternative function when the object is not supported
}
