% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.R
\name{rtolatex}
\alias{rtolatex}
\title{Turn R output into Latex}
\usage{
rtolatex(x, ...)
}
\arguments{
\item{x}{The R object to be converted}

\item{...}{Additional object to the converted}
}
\value{
Latex inputs
}
\description{
Turn R output into Latex
}
