% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.R
\name{rtolatex_error}
\alias{rtolatex_error}
\title{Default method for unsupported classes}
\usage{
rtolatex_error(x, ...)
}
\arguments{
\item{x}{An unsupported R object.}

\item{...}{Ignored.}
}
\description{
Default method for unsupported classes
}
