% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.data.frame.R
\name{rtolatex.data.frame}
\alias{rtolatex.data.frame}
\title{convert dataframe into latex}
\usage{
\method{rtolatex}{data.frame}(x, caption = "", label = "", digits = 2, ...)
}
\arguments{
\item{x}{a dataframe}

\item{caption}{table caption}

\item{label}{label for table}

\item{digits}{digits to be rounded to}

\item{...}{ignored only for syntax}
}
\value{
latex snippet
}
\description{
transforms dataframe to a latex tabular form
special characters are automatically adjust to latex
}
\examples{
rtolatex(head(mtcars), caption = "Motor Trend Car Data", label = "tab:mtcars")
}
