% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.data.frame.R
\name{rtolatex.data.frame}
\alias{rtolatex.data.frame}
\title{convert dataframe into latex}
\usage{
\method{rtolatex}{data.frame}(x, caption = "", label = "", ...)
}
\arguments{
\item{x}{a dataframe}

\item{caption}{table caption}

\item{label}{label for table}

\item{...}{ignored only for syntax}
}
\value{
latex snippet
}
\description{
Takes a data frame and turns it into a LaTeX tabular environment.
Special characters in the data are automatically escaped.
}
\examples{
rtolatex(head(mtcars), caption = "Motor Trend Car Data", label = "tab:mtcars")
}
