% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.ggplot.R
\name{rtolatex.ggplot}
\alias{rtolatex.ggplot}
\title{convert ggplot to latex}
\usage{
\method{rtolatex}{ggplot}(x, filename, path = "outputs", width = 6, height = 4, ...)
}
\arguments{
\item{x}{the ggplot}

\item{filename}{name of the file}

\item{path}{folder path}

\item{width}{width of the image}

\item{height}{height of the image}

\item{...}{ignored only for syntax}
}
\value{
latex snippet
}
\description{
saves ggplots as pdf files
includes graphics command to insert it into your document
}
\examples{
library(ggplot2)
p <- ggplot(data.frame(x = rnorm(30)), aes(x = x)) + geom_histogram(bins = 10)
rtolatex(p, filename = "my_plot")
}
