% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.htest.R
\name{rtolatex.htest}
\alias{rtolatex.htest}
\title{convert a hypothesis test results to latex}
\usage{
\method{rtolatex}{htest}(x, ...)
}
\arguments{
\item{x}{result of a hypothesis test}

\item{...}{ignored only for syntax}
}
\value{
latex snippet
}
\description{
Takes the output of a hypothesis test like t.test() or chisq.test()
and formats it as a ready to use inline sentence in APA style.
}
\examples{
rtolatex(t.test(rnorm(30), rnorm(30)))
rtolatex(chisq.test(c(10, 20, 30)))
}
