% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.numeric.R
\name{rtolatex.numeric}
\alias{rtolatex.numeric}
\title{convert numbers to latex}
\usage{
\method{rtolatex}{numeric}(x, name, digits = 2, ...)
}
\arguments{
\item{x}{a single numeric value}

\item{name}{name of latex command}

\item{digits}{decimal place rounding}

\item{...}{ignored only for syntax}
}
\value{
latex snippet
}
\description{
turns the number to a latex reference
automatic updates when results are refreshed
}
\examples{
rtolatex(0.94, name = "accuracy", digits = 2)
}
