% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtolatex.R
\name{rtolatex}
\alias{rtolatex}
\title{turn R output into latex}
\usage{
rtolatex(x, ...)
}
\arguments{
\item{x}{object to be converted into latex}

\item{...}{other objects that need to be converted}
}
\value{
latex formatted inputs
}
\description{
turn R output into latex
}
