% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_textfile.R
\name{write_textfile}
\alias{write_textfile}
\title{writing all snippets as .tex files for latex}
\usage{
write_textfile(doc, ...)
}
\arguments{
\item{doc}{document container object}

\item{...}{ignored only for syntax}
}
\value{
returns document container
}
\description{
creates a .tex file for each entry in the document container
as a separate .tex file in the output folder
files get reference in latex using the input command
}
\examples{
doc <- document_container()
doc <- add_snippet(doc, rtolatex(0.94, name = "accuracy"), "accuracy")
write_textfile(doc)
}
